send DirEntries to output instead of just path
